% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_public}
\alias{data_public}
\title{Example Data}
\format{
`data_public` is example data frame with 10 rows and 511 columns
}
\source{
<https://comsamozambique.org/data-access>
}
\usage{
data_public
}
\description{
A subset of mortaility surveilance data from the Countrywide Mortality Surveillance for Action project
}
\keyword{datasets}
