% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk2EAVA.R
\name{odk2EAVA}
\alias{odk2EAVA}
\title{odk2EAVA}
\source{
: odk2EAVA builds on top of odk2openVA_v151() from the CrossVA package https://github.com/cran/CrossVA
}
\usage{
odk2EAVA(odk, id_col)
}
\arguments{
\item{odk}{A data frame which used open data kit (odk) to obtain 2016 WHO VA questionnaire responses}

\item{id_col}{A unique identifier for each record within the odk data frame}
}
\value{
A data frame in openVA format, where each row is a death, and each column corresponds to a
standardized openVA symptom or attribute variable as defined in ccva_codebook

openVA variables are derived from the 2016 WHO VA instrument using Peter Byass's InterVA-5 user guide.
Further details on the coding and underlying methodology are available from
http://www.byass.uk/interva/products.html
Example codebook mapping between openVA variable names and their original questionnaire
question is shown below. See ccva_codebook for comprehensive mapping:
\describe{
  \item{i004a}{Did the death occur during the wet season?}
  \item{i004b}{Did the death occur during the dry season?}
  \item{i019a}{Male sex indicator}
  \item{i019b}{Female sex indicator}
  \item{i174o}{Chest pain}
  \item{i181o}{Diarrhoea}
  \item{i363o}{Low birth weight (<2.5 kg)}
}
Variables ending in 'o' are coded as binary indicators (1 = yes, 0 = no, NA = missing),
Variables ending in 'a' are non-binary indicators such as symptom duration in days or weeks
}
\description{
Converts 2016 WHO verbal autopsy (VA) data to an input file for Expert Algorithm Verbal Autopsy cause of death assignment by the codEAVA() function
}
\examples{
{
# load embedded example data or data from WHO 2016 Verbal Autopsy Questionnaire
data <- as.data.frame(data_public)
# run odk2EAVA()
output <- odk2EAVA(data, id_col  = "comsa_id")
# view data converted for use in codEAVA()
head(output)
}
}
\references{
Thomas J, Choi E, Li Z, Maire N, McCormick T, Byass P, Clark S (2021). CrossVA: Verbal Autopsy Data Transformation for InSilicoVA and InterVA5 Algorithms_. R package version 1.0.0, <https://CRAN.R-project.org/package=CrossVA>.
}
