% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codEAVA.R
\name{codEAVA}
\alias{codEAVA}
\title{codEAVA}
\usage{
codEAVA(df, age_group)
}
\arguments{
\item{df}{A data frame frame with 2016 WHO VA responses in openVA input format}

\item{age_group}{Age group input, either "neonate" or "child"}
}
\value{
A two-column data frame with unique identifier and cause of death
}
\description{
Assigns cause of death by Expert Algorithm Verbal Autopsy
}
