% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk2EAVA.R
\name{odk2EAVA}
\alias{odk2EAVA}
\title{odk2EAVA}
\usage{
data_public is de-identified example data which is available publicly
}
\arguments{
\item{odk}{A data frame frame which used open data kit (odk) to obtain 2016 WHO VA questionnaire responses}

\item{id_col}{A unique identifier for each record within the odk data frame}
}
\value{
A data frame that contains variable names and values which have been converted to openVA convention
}
\description{
Converts 2016 WHO verbal autopsy (VA) data to an input file for Expert Algorithm Verbal Autopsy cause of death assignment by the codEAVA() function
}
